Don't free the cmd if it's [p]subscribe
If one sent a [p]subscribe command through a websocket, the cmd
struct got freed after the first response was received.  Subscribe
commands generate multiple responses, so this resulted in accessing
already-freed data, and double-freeing.

This patch adds a check to see if it's a subscribe command, and if it
is avoids the freeing.