Use msgpack_pack_v4raw(_body) with new msgpack-c versions
In msgpack-c's 1.0.0 release, the code was changed to be compatible with
the v5 msgpack spec, specifically separating out the STR and BIN types
as replacements for the old RAW type.

While the STR type is likely the right replacement type to use, it does
introduce a str 8 variant that wasn't present for the old RAW type.  For
better compatibility, the msgpack_pack_v4raw functions were added to
directly map to the old functionality.

Signed-off-by: James McCoy <jamessan@jamessan.com>