Only process cx: close header if full request was read (fixes #194)
The presence of a `Connection: close` header sets a flag on the request,
which we process by no longer running the HTTP parser as soon as it
returns from parsing a chunk. Since HTTP 100 uploads require at least 2
reads, we need to handle the `Connection: close` only if we've reached
the end of the request. This change sets a flag in `on_message_complete`,
which we use in combination with the header to stop reading.

Also adding a test to reproduce this issue and including it in CI job.